backend/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.11.1
attrs==26.1.0
billiard==4.2.3
branca==0.8.2
cachetools==5.5.2
celery==5.5.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.6
cycler==0.12.1
distro==1.9.0
dj-database-url==2.3.0
Django==6.0.3
django-cors-headers==4.9.0
django-environ==0.12.0
django-filter==25.2
django-redis==6.0.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
drf-yasg==1.21.11
et_xmlfile==2.0.0
eval_type_backport==0.3.1
folium==0.20.0
fonttools==4.60.1
google-auth==2.49.1
google-auth-oauthlib==1.2.1
google-genai==1.69.0
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.1
inflection==0.5.1
iniconfig==2.3.0
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
jsonpath-python==1.1.5
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
kombu==5.5.4
MarkupSafe==3.0.3
matplotlib==3.10.7
mistralai==2.1.3
mysqlclient==2.2.8
numpy==2.3.5
oauthlib==3.3.1
openai==2.8.1
openpyxl==3.1.5
opentelemetry-api==1.39.1
opentelemetry-semantic-conventions==0.60b1
packaging==25.0
pandas==2.3.3
pillow==12.1.1
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
PyJWT==2.12.1
pyparsing==3.2.5
pytest==9.0.2
pytest-django==4.12.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
qrcode==8.2
redis==7.1.0
referencing==0.37.0
regex==2025.11.3
requests==2.33.0
requests-oauthlib==2.0.0
rpds-py==0.30.0
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.16.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
tenacity==9.1.4
threadpoolctl==3.6.0
tiktoken==0.12.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.2.14
websockets==16.0
whitenoise==6.11.0
xyzservices==2025.11.0
zipp==3.23.0
stripe==11.4.0
firebase-admin==6.6.0
python-dotenv==1.2.1
cloudinary==1.41.0
django-cloudinary-storage==0.3.0
